# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST and gRPC API for managing users and outbox events",
        "title": "User Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/notes": {
            "get": {
                "description": "Get a list of all notes with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Get all notes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/notesservice_internal_models_domain.Note"
                            }
                        }
                    },
                    "500": {
                        "description": "failed to get notes",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing note record",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Update note",
                "parameters": [
                    {
                        "description": "Updated note data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notesservice_internal_models_swagger.NoteRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to update note",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Insert a new note record",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Create note",
                "parameters": [
                    {
                        "description": "Inserted note data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notesservice_internal_models_swagger.NoteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "note already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to insert note",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a note record by ID in body",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Delete note",
                "parameters": [
                    {
                        "description": "ID container: {'id': 'uuid'}",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notesservice_internal_models_swagger.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to delete note",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notes/by-user": {
            "get": {
                "description": "Get a list of notes for a specific user via query parameter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Get notes by user",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User UUID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/notesservice_internal_models_domain.Note"
                            }
                        }
                    },
                    "400": {
                        "description": "missing or invalid id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to get notes by user",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notes/detail": {
            "get": {
                "description": "Get details of a single note by its UUID in query",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Get note details",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Note UUID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notesservice_internal_models_domain.Note"
                        }
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "note not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to get note by id",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "notesservice_internal_models_domain.Note": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Молоко, сыр, хлеб"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-02-05T13:40:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "is_private": {
                    "type": "boolean",
                    "example": true
                },
                "title": {
                    "type": "string",
                    "example": "Купить продукты"
                },
                "user_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
        "notesservice_internal_models_swagger.DeleteRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "notesservice_internal_models_swagger.NoteRequest": {
            "type": "object",
            "properties": {
                "note": {
                    "$ref": "#/definitions/notesservice_internal_models_domain.Note"
                }
            }
        }
    }
}